fix(dma): merge ping-pong RX stream from refactor/366 branch
Bring in the ping-pong TAILDESC advancement architecture:
- ISR clears CMPLT + advances TAILDESC, DMA never stops
- Software coalescing with hw_irq_threshold capped at 32
- irq_threshold widened to uint16_t
- DLY_IRQ permanently disabled for RX stream
- Default NUM_RX_BD increased to 1024

Hardware validated at 200 Hz and 32 kHz continuous capture with
subscribe/data/stop/restart cycles.

+++ include/zephyr/drivers/dma/dma_xlnx_axi_dma_sg.h
@@ -40,7 +40,7 @@ typedef void (*dma_xlnx_sg_rx_stream_cb_t)(const struct device *dev, void *user_
  */
 struct dma_xlnx_sg_rx_stream_cfg {
 	uint32_t bd_bytes;
-	uint8_t irq_threshold;
+	uint16_t irq_threshold;
 	dma_xlnx_sg_rx_stream_cb_t callback;
 	void *user_data;
 };
